Made some improvements on the n_style feed forward network. it now trains on only the scale and offset variable for the second style onward. This greatly reduces training time.
Another improvement is: it now supports real time rendering. I don't have a camera yet so I can't do camera rendering. But this is really exciting. The result was not so satisfactory though. The color is still a little wierd. Probably the number of training round is too little. All the papers uses like 40000+ rounds of training.